Finalized Update to Printing
Completely refactored the printing code to rely on a `TablePrinter`
type, which encodes settings.

Now the code abstracts away from the specifics of the latex/ascii
printing into a number of methods that print each line one at a time.
Things like the separator character, the amount of whitespace padding,
where the standard errors appear, etc... are all controlled by
parameters given to the `TablePrinter` object.