Twoway Tabulate (#17)
* First pass at two-way tabulate function

I don't like the way it handles the labels.  Need to implement row
labels as a new feature before merging this into master.

* First pass at two-way tabulate function

I don't like the way it handles the labels.  Need to implement row
labels as a new feature before merging this into master.

* Fix Code for 2 Way Tables

You can't convert a dataframe to an array anymore...

* Fixed a bug in tabulate

when the field in question was `:N`, tabulate gave an incorrect result.

* Added Tests!

* Fix Error in Twoway Tabulations

If one of the fields only appears for a subset of the others, we were
calculating the tabulations as zero.  But we actually needed to
calculate the column totals replacing missing as 0.

Need to add an extra test.

* Fix Performance Problem in Summarize

* Small Fixes